Cleanup API gateway and resource helpers
