+++ xtrinode/controllers/nodepool_aws_managed.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 
 	analyticsv1 "github.com/xtrinode/xtrinode/api/v1"
 )
@@ -20,16 +19,7 @@ func (n *NodePoolAdapter) ensureAWSManagedMachinePool(ctx context.Context, xtrin
 	// Get defaults
 	defaults := getNodePoolDefaults(nodePool, xtrinode)
 
-	// Check if MachinePool already exists to gate replica setting
-	existingCheck := &unstructured.Unstructured{}
-	existingCheck.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "cluster.x-k8s.io",
-		Version: "v1beta1",
-		Kind:    "MachinePool",
-	})
-	existingCheck.SetName(poolName)
-	existingCheck.SetNamespace(xtrinode.Namespace)
-	resourceExists, err := checkResourceExists(n.client, ctx, existingCheck)
+	resourceExists, err := managedMachinePoolExists(n.client, ctx, poolName, xtrinode.Namespace)
 	if err != nil {
 		return fmt.Errorf("failed to check if MachinePool exists: %w", err)
 	}
@@ -40,25 +30,8 @@ func (n *NodePoolAdapter) ensureAWSManagedMachinePool(ctx context.Context, xtrin
 	}
 
 	// Step 1: Create/update AWSManagedMachinePool (provider infra CRD)
-	infraPool := &unstructured.Unstructured{}
-	infraPool.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "infrastructure.cluster.x-k8s.io",
-		Version: "v1beta2",
-		Kind:    "AWSManagedMachinePool",
-	})
-	infraPool.SetName(poolName)
-	infraPool.SetNamespace(xtrinode.Namespace)
-
-	// Set cluster label (CAPI standard for cluster association)
-	labels := infraPool.GetLabels()
-	if labels == nil {
-		labels = make(map[string]string)
-	}
-	labels["cluster.x-k8s.io/cluster-name"] = clusterName
-	infraPool.SetLabels(labels)
-
-	// CAPI MachinePool must become the controller owner for managed infra pools.
-	if err := setXTrinodeNonControllerOwnerReference(infraPool, xtrinode); err != nil {
+	infraPool, err := newManagedInfrastructurePool("aws", poolName, xtrinode.Namespace, clusterName, xtrinode)
+	if err != nil {
 		return err
 	}
